docs: Improve backticks for example
Instead of having a backslash and the prompt to '(replace final line with triple ` )', this formatting renders the example code in the desired way. Check it out in the 'Preview' on GitHub.

Also, if one uses the 'Copy' button in the top-right corner, they'll get the desired code with the correct final line.